backend/requirements.txt
Flask  
flask-cors
datetime 
uuid
Flask-SQLAlchemy
PyJWT
psycopg2
langchain
langchain_cohere
langgraph
yfinance
faiss-cpu
bcrypt
